+++ rust/crd/src/affinity.rs
@@ -1,0 +1,189 @@
+use stackable_operator::{
+    commons::affinity::{affinity_between_role_pods, StackableAffinityFragment},
+    k8s_openapi::api::core::v1::PodAntiAffinity,
+};
+
+use crate::{SupersetRole, APP_NAME};
+
+pub fn get_affinity(cluster_name: &str, role: &SupersetRole) -> StackableAffinityFragment {
+    StackableAffinityFragment {
+        pod_affinity: None,
+        pod_anti_affinity: Some(PodAntiAffinity {
+            preferred_during_scheduling_ignored_during_execution: Some(vec![
+                affinity_between_role_pods(APP_NAME, cluster_name, &role.to_string(), 70),
+            ]),
+            required_during_scheduling_ignored_during_execution: None,
+        }),
+        node_affinity: None,
+        node_selector: None,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use std::collections::BTreeMap;
+
+    use crate::SupersetCluster;
+    use stackable_operator::{
+        commons::affinity::{StackableAffinity, StackableNodeSelector},
+        k8s_openapi::{
+            api::core::v1::{
+                NodeAffinity, NodeSelector, NodeSelectorRequirement, NodeSelectorTerm,
+                PodAffinityTerm, PodAntiAffinity, WeightedPodAffinityTerm,
+            },
+            apimachinery::pkg::apis::meta::v1::LabelSelector,
+        },
+    };
+
+    #[test]
+    fn test_affinity_defaults() {
+        let input = r#"
+        apiVersion: superset.stackable.tech/v1alpha1
+        kind: SupersetCluster
+        metadata:
+          name: simple-superset
+        spec:
+          image:
+            productVersion: 1.5.1
+            stackableVersion: "23.1"
+          credentialsSecret: simple-superset-credentials
+          nodes:
+            roleGroups:
+              default:
+                replicas: 1
+        "#;
+        let superset: SupersetCluster = serde_yaml::from_str(input).expect("illegal test input");
+        let merged_config = superset
+            .merged_config(&SupersetRole::Node, &superset.node_rolegroup_ref("default"))
+            .unwrap();
+
+        assert_eq!(
+            merged_config.affinity,
+            StackableAffinity {
+                pod_affinity: None,
+                pod_anti_affinity: Some(PodAntiAffinity {
+                    preferred_during_scheduling_ignored_during_execution: Some(vec![
+                        WeightedPodAffinityTerm {
+                            pod_affinity_term: PodAffinityTerm {
+                                label_selector: Some(LabelSelector {
+                                    match_expressions: None,
+                                    match_labels: Some(BTreeMap::from([
+                                        (
+                                            "app.kubernetes.io/name".to_string(),
+                                            "superset".to_string(),
+                                        ),
+                                        (
+                                            "app.kubernetes.io/instance".to_string(),
+                                            "simple-superset".to_string(),
+                                        ),
+                                        (
+                                            "app.kubernetes.io/component".to_string(),
+                                            "node".to_string(),
+                                        )
+                                    ]))
+                                }),
+                                namespace_selector: None,
+                                namespaces: None,
+                                topology_key: "kubernetes.io/hostname".to_string(),
+                            },
+                            weight: 70
+                        }
+                    ]),
+                    required_during_scheduling_ignored_during_execution: None,
+                }),
+                node_affinity: None,
+                node_selector: None,
+            }
+        );
+    }
+
+    #[test]
+    fn test_affinity_legacy_node_selector() {
+        let input = r#"
+        apiVersion: superset.stackable.tech/v1alpha1
+        kind: SupersetCluster
+        metadata:
+          name: simple-superset
+        spec:
+          image:
+            productVersion: 1.5.1
+            stackableVersion: "23.1"
+          credentialsSecret: simple-superset-credentials
+          nodes:
+            roleGroups:
+              default:
+                replicas: 1
+                selector:
+                  matchLabels:
+                    disktype: ssd
+                  matchExpressions:
+                    - key: topology.kubernetes.io/zone
+                      operator: In
+                      values:
+                        - antarctica-east1
+                        - antarctica-west1
+        "#;
+        let superset: SupersetCluster = serde_yaml::from_str(input).expect("illegal test input");
+        let merged_config = superset
+            .merged_config(&SupersetRole::Node, &superset.node_rolegroup_ref("default"))
+            .unwrap();
+
+        assert_eq!(
+            merged_config.affinity,
+            StackableAffinity {
+                pod_affinity: None,
+                pod_anti_affinity: Some(PodAntiAffinity {
+                    preferred_during_scheduling_ignored_during_execution: Some(vec![
+                        WeightedPodAffinityTerm {
+                            pod_affinity_term: PodAffinityTerm {
+                                label_selector: Some(LabelSelector {
+                                    match_expressions: None,
+                                    match_labels: Some(BTreeMap::from([
+                                        (
+                                            "app.kubernetes.io/name".to_string(),
+                                            "superset".to_string(),
+                                        ),
+                                        (
+                                            "app.kubernetes.io/instance".to_string(),
+                                            "simple-superset".to_string(),
+                                        ),
+                                        (
+                                            "app.kubernetes.io/component".to_string(),
+                                            "node".to_string(),
+                                        )
+                                    ]))
+                                }),
+                                namespace_selector: None,
+                                namespaces: None,
+                                topology_key: "kubernetes.io/hostname".to_string(),
+                            },
+                            weight: 70
+                        }
+                    ]),
+                    required_during_scheduling_ignored_during_execution: None,
+                }),
+                node_affinity: Some(NodeAffinity {
+                    preferred_during_scheduling_ignored_during_execution: None,
+                    required_during_scheduling_ignored_during_execution: Some(NodeSelector {
+                        node_selector_terms: vec![NodeSelectorTerm {
+                            match_expressions: Some(vec![NodeSelectorRequirement {
+                                key: "topology.kubernetes.io/zone".to_string(),
+                                operator: "In".to_string(),
+                                values: Some(vec![
+                                    "antarctica-east1".to_string(),
+                                    "antarctica-west1".to_string()
+                                ]),
+                            }]),
+                            match_fields: None,
+                        }]
+                    }),
+                }),
+                node_selector: Some(StackableNodeSelector {
+                    node_selector: BTreeMap::from([("disktype".to_string(), "ssd".to_string())])
+                }),
+            }
+        );
+    }
+}